requirements.txt
pdfminer
ebooklib
beautifulsoup4
python-docx
pandas
tqdm
configparser
google-generativeai